[SCMOD-9351] Change group and package
