+++ fortify-issue-manager-cli/pom.xml
@@ -22,7 +22,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>com.microfocus.sepg.fortify</groupId>
+        <groupId>com.microfocus.security.automation.fortify</groupId>
         <artifactId>fortify-issue-manager-aggregator</artifactId>
         <version>1.0.0-SNAPSHOT</version>
     </parent>
@@ -31,7 +31,7 @@
 
     <dependencies>
         <dependency>
-            <groupId>com.microfocus.sepg.fortify</groupId>
+            <groupId>com.microfocus.security.automation.fortify</groupId>
             <artifactId>fortify-issue-manager</artifactId>
         </dependency>
         <dependency>
@@ -58,7 +58,7 @@
                     <archive>
                         <manifest>
                             <addClasspath>true</addClasspath>
-                            <mainClass>com.microfocus.sepg.fortify.issue.manager.cli.Program</mainClass>
+                            <mainClass>com.microfocus.security.automation.fortify.issue.manager.cli.Program</mainClass>
                         </manifest>
                     </archive>
                 </configuration>
