+++ fortify-issue-manager/pom.xml
@@ -22,7 +22,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>com.microfocus.sepg.fortify</groupId>
+        <groupId>com.microfocus.security.automation.fortify</groupId>
         <artifactId>fortify-issue-manager-aggregator</artifactId>
         <version>1.0.0-SNAPSHOT</version>
     </parent>
