+++ fortify-issue-manager/src/main/java/com/microfocus/security/automation/fortify/issue/manager/models/Vulnerability.java
@@ -1,0 +1,92 @@
+/*
+ * Copyright 2020 Micro Focus or one of its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.microfocus.security.automation.fortify.issue.manager.models;
+
+public final class Vulnerability
+{
+    private int id;
+    private int releaseId;
+    private int severity;
+    private String severityString;
+    private String category;
+    private String vulnId;
+    private String status;
+    private String bugSubmitted;
+    private String bugLink;
+    private String primaryLocation;
+    private String lineNumber;
+    private String checkId;
+
+    public int getId()
+    {
+        return id;
+    }
+
+    public int getReleaseId()
+    {
+        return releaseId;
+    }
+
+    public int getSeverity()
+    {
+        return severity;
+    }
+
+    public String getSeverityString()
+    {
+        return severityString;
+    }
+
+    public String getCategory()
+    {
+        return category;
+    }
+
+    public String getVulnId()
+    {
+        return vulnId;
+    }
+
+    public String getStatus()
+    {
+        return status;
+    }
+
+    public String getBugSubmitted()
+    {
+        return bugSubmitted;
+    }
+
+    public String getBugLink()
+    {
+        return bugLink;
+    }
+
+    public String getPrimaryLocation()
+    {
+        return primaryLocation;
+    }
+
+    public String getLineNumber()
+    {
+        return lineNumber;
+    }
+
+    public String getCheckId()
+    {
+        return checkId;
+    }
+}
