--- fortify-issue-manager/src/main/java/com/microfocus/sepg/fortify/issue/manager/models/Category.java
@@ -1,81 +1,0 @@
-/*
- * Copyright 2020 Micro Focus or one of its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.microfocus.sepg.fortify.issue.manager.models;
-
-public final class Category
-{
-    private final String name;
-    private final int severity;
-
-    public Category(String name, int severity)
-    {
-        this.name = name;
-        this.severity = severity;
-    }
-
-    public String getName()
-    {
-        return name;
-    }
-
-    public int getSeverity()
-    {
-        return severity;
-    }
-
-    @Override
-    public int hashCode()
-    {
-        final int prime = 31;
-        int result = 1;
-        result = prime * result + ((name == null) ? 0 : name.hashCode());
-        result = prime * result + severity;
-        return result;
-    }
-
-    @Override
-    public boolean equals(Object obj)
-    {
-        if (this == obj)
-            return true;
-        if (obj == null)
-            return false;
-        if (getClass() != obj.getClass())
-            return false;
-        Category other = (Category) obj;
-        if (name == null)
-        {
-            if (other.name != null)
-                return false;
-        } else if (!name.equals(other.name))
-            return false;
-        if (severity != other.severity)
-            return false;
-        return true;
-    }
-
-    @Override
-    public String toString()
-    {
-        StringBuilder builder = new StringBuilder();
-        builder.append("Category [");
-        builder.append(name);
-        builder.append(", ");
-        builder.append(severity);
-        builder.append("]");
-        return builder.toString();
-    }
-}
